requirements.txt

fastapi
uvicorn
python-multipart
aiofiles
pypdf
sentence-transformers
pinecone-client
sqlalchemy
aiosqlite  
redis
aioredis
pydantic
python-dotenv